refactor: Переименовать input на textarea, проводить рефакторинг компонента
