+++ src/components/ui/text-field/index.ts
@@ -1,1 +1,1 @@
-export * from '../text-field/text-field';
+export * from './text-field';
