+++ src/components/ui/text-field/input.stories.tsx
@@ -1,6 +1,6 @@
 import {ComponentStory, ComponentMeta} from '@storybook/react';
 
-import Input from './input';
+import {Input} from './text-field';
 
 export default {
   title: 'UI/Input',
@@ -19,14 +19,17 @@ const Template: ComponentStory<typeof Input> = (args) => <Input {...args} />;
 export const Default = Template.bind({});
 Default.args = {
   type: 'text',
+  placeholder: 'Введите имя'
 };
 
 export const Email = Template.bind({});
 Email.args = {
   type: 'e-mail',
+  placeholder: 'Введите e-mail'
 };
 
 export const Password = Template.bind({});
 Password.args = {
   type: 'password',
+  placeholder: 'Введите пароль'
 };
